Enhance DevQL documentation and CLI commands for improved clarity and support for multiple backends. Added command support matrix, updated command descriptions, and improved database interaction for DuckDB and SQLite. Refactored ingestion and query execution to accommodate relational storage, ensuring compatibility with both Postgres and SQLite. Introduced new schema initialization for DuckDB and enhanced error handling for database operations.
