+++ bitloops_cli/src/engine/devql/ingestion/schema/relational_sqlite_schema.rs
@@ -1,0 +1,214 @@
+fn sqlite_schema_sql() -> &'static str {
+    r#"
+CREATE TABLE IF NOT EXISTS repositories (
+    repo_id TEXT PRIMARY KEY,
+    provider TEXT NOT NULL,
+    organization TEXT NOT NULL,
+    name TEXT NOT NULL,
+    default_branch TEXT,
+    created_at TEXT DEFAULT (datetime('now'))
+);
+
+CREATE TABLE IF NOT EXISTS commits (
+    commit_sha TEXT PRIMARY KEY,
+    repo_id TEXT NOT NULL,
+    author_name TEXT,
+    author_email TEXT,
+    commit_message TEXT,
+    committed_at TEXT
+);
+
+CREATE TABLE IF NOT EXISTS file_state (
+    repo_id TEXT NOT NULL,
+    commit_sha TEXT NOT NULL,
+    path TEXT NOT NULL,
+    blob_sha TEXT NOT NULL,
+    PRIMARY KEY (repo_id, commit_sha, path)
+);
+
+CREATE INDEX IF NOT EXISTS file_state_blob_idx
+ON file_state (repo_id, blob_sha);
+
+CREATE INDEX IF NOT EXISTS file_state_commit_idx
+ON file_state (repo_id, commit_sha);
+
+CREATE TABLE IF NOT EXISTS current_file_state (
+    repo_id TEXT NOT NULL,
+    path TEXT NOT NULL,
+    commit_sha TEXT NOT NULL,
+    blob_sha TEXT NOT NULL,
+    committed_at TEXT NOT NULL,
+    updated_at TEXT DEFAULT (datetime('now')),
+    PRIMARY KEY (repo_id, path)
+);
+
+CREATE TABLE IF NOT EXISTS artefacts (
+    artefact_id TEXT PRIMARY KEY,
+    symbol_id TEXT,
+    repo_id TEXT NOT NULL,
+    blob_sha TEXT NOT NULL,
+    path TEXT NOT NULL,
+    language TEXT NOT NULL,
+    canonical_kind TEXT,
+    language_kind TEXT,
+    symbol_fqn TEXT,
+    parent_artefact_id TEXT,
+    start_line INTEGER NOT NULL,
+    end_line INTEGER NOT NULL,
+    start_byte INTEGER NOT NULL,
+    end_byte INTEGER NOT NULL,
+    signature TEXT,
+    modifiers TEXT NOT NULL DEFAULT '[]',
+    docstring TEXT,
+    content_hash TEXT,
+    created_at TEXT DEFAULT (datetime('now'))
+);
+
+CREATE INDEX IF NOT EXISTS artefacts_blob_idx
+ON artefacts (repo_id, blob_sha);
+
+CREATE INDEX IF NOT EXISTS artefacts_path_idx
+ON artefacts (repo_id, path);
+
+CREATE INDEX IF NOT EXISTS artefacts_kind_idx
+ON artefacts (repo_id, canonical_kind);
+
+CREATE INDEX IF NOT EXISTS artefacts_symbol_idx
+ON artefacts (repo_id, symbol_id);
+
+CREATE TABLE IF NOT EXISTS artefacts_current (
+    repo_id TEXT NOT NULL,
+    symbol_id TEXT NOT NULL,
+    artefact_id TEXT NOT NULL,
+    commit_sha TEXT NOT NULL,
+    blob_sha TEXT NOT NULL,
+    path TEXT NOT NULL,
+    language TEXT NOT NULL,
+    canonical_kind TEXT,
+    language_kind TEXT,
+    symbol_fqn TEXT,
+    parent_symbol_id TEXT,
+    parent_artefact_id TEXT,
+    start_line INTEGER NOT NULL,
+    end_line INTEGER NOT NULL,
+    start_byte INTEGER NOT NULL,
+    end_byte INTEGER NOT NULL,
+    signature TEXT,
+    modifiers TEXT NOT NULL DEFAULT '[]',
+    docstring TEXT,
+    content_hash TEXT,
+    updated_at TEXT DEFAULT (datetime('now')),
+    PRIMARY KEY (repo_id, symbol_id)
+);
+
+CREATE INDEX IF NOT EXISTS artefacts_current_path_idx
+ON artefacts_current (repo_id, path);
+
+CREATE INDEX IF NOT EXISTS artefacts_current_kind_idx
+ON artefacts_current (repo_id, canonical_kind);
+
+CREATE INDEX IF NOT EXISTS artefacts_current_artefact_idx
+ON artefacts_current (repo_id, artefact_id);
+
+CREATE INDEX IF NOT EXISTS artefacts_current_symbol_fqn_idx
+ON artefacts_current (repo_id, symbol_fqn);
+
+CREATE TABLE IF NOT EXISTS artefact_edges (
+    edge_id TEXT PRIMARY KEY,
+    repo_id TEXT NOT NULL,
+    blob_sha TEXT NOT NULL,
+    from_artefact_id TEXT NOT NULL,
+    to_artefact_id TEXT,
+    to_symbol_ref TEXT,
+    edge_kind TEXT NOT NULL,
+    language TEXT NOT NULL,
+    start_line INTEGER,
+    end_line INTEGER,
+    metadata TEXT DEFAULT '{}',
+    created_at TEXT DEFAULT (datetime('now')),
+    CHECK (to_artefact_id IS NOT NULL OR to_symbol_ref IS NOT NULL),
+    CHECK (
+        (start_line IS NULL AND end_line IS NULL)
+        OR (start_line IS NOT NULL AND end_line IS NOT NULL AND start_line > 0 AND end_line >= start_line)
+    )
+);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_blob_idx
+ON artefact_edges (repo_id, blob_sha);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_from_idx
+ON artefact_edges (repo_id, from_artefact_id, edge_kind);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_to_idx
+ON artefact_edges (repo_id, to_artefact_id, edge_kind);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_kind_idx
+ON artefact_edges (repo_id, edge_kind);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_symbol_ref_idx
+ON artefact_edges (repo_id, edge_kind, to_symbol_ref);
+
+CREATE UNIQUE INDEX IF NOT EXISTS artefact_edges_natural_uq
+ON artefact_edges (
+    repo_id,
+    blob_sha,
+    from_artefact_id,
+    edge_kind,
+    COALESCE(to_artefact_id, ''),
+    COALESCE(to_symbol_ref, ''),
+    COALESCE(start_line, -1),
+    COALESCE(end_line, -1)
+);
+
+CREATE TABLE IF NOT EXISTS artefact_edges_current (
+    edge_id TEXT PRIMARY KEY,
+    repo_id TEXT NOT NULL,
+    commit_sha TEXT NOT NULL,
+    blob_sha TEXT NOT NULL,
+    path TEXT NOT NULL,
+    from_symbol_id TEXT NOT NULL,
+    from_artefact_id TEXT NOT NULL,
+    to_symbol_id TEXT,
+    to_artefact_id TEXT,
+    to_symbol_ref TEXT,
+    edge_kind TEXT NOT NULL,
+    language TEXT NOT NULL,
+    start_line INTEGER,
+    end_line INTEGER,
+    metadata TEXT DEFAULT '{}',
+    updated_at TEXT DEFAULT (datetime('now')),
+    CHECK (to_symbol_id IS NOT NULL OR to_symbol_ref IS NOT NULL),
+    CHECK (
+        (start_line IS NULL AND end_line IS NULL)
+        OR (start_line IS NOT NULL AND end_line IS NOT NULL AND start_line > 0 AND end_line >= start_line)
+    )
+);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_current_path_idx
+ON artefact_edges_current (repo_id, path);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_current_from_idx
+ON artefact_edges_current (repo_id, from_symbol_id, edge_kind);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_current_to_idx
+ON artefact_edges_current (repo_id, to_symbol_id, edge_kind);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_current_kind_idx
+ON artefact_edges_current (repo_id, edge_kind);
+
+CREATE INDEX IF NOT EXISTS artefact_edges_current_symbol_ref_idx
+ON artefact_edges_current (repo_id, to_symbol_ref);
+
+CREATE UNIQUE INDEX IF NOT EXISTS artefact_edges_current_natural_uq
+ON artefact_edges_current (
+    repo_id,
+    from_symbol_id,
+    edge_kind,
+    COALESCE(to_symbol_id, ''),
+    COALESCE(to_symbol_ref, ''),
+    COALESCE(start_line, -1),
+    COALESCE(end_line, -1),
+    COALESCE(metadata, '{}')
+);
+"#
+}
