+++ bitloops_cli/src/engine/devql/tests/devql_tests/core_and_ingestion.rs
@@ -1,0 +1,302 @@
+#[test]
+fn sql_helpers_escape_nullable_and_json_values() {
+    assert_eq!(sql_nullable_text(None), "NULL");
+    assert_eq!(sql_nullable_text(Some("O'Reilly")), "'O''Reilly'");
+    assert_eq!(
+        sql_jsonb_text_array(&["O'Reilly".to_string(), "plain".to_string()]),
+        r#"'["O''Reilly","plain"]'::jsonb"#
+    );
+}
+
+#[test]
+fn supported_symbol_languages_are_whitelisted() {
+    for language in ["typescript", "javascript", "rust"] {
+        assert!(
+            is_supported_symbol_language(language),
+            "{language} should be supported"
+        );
+    }
+
+    for language in ["python", "go", ""] {
+        assert!(
+            !is_supported_symbol_language(language),
+            "{language} should not be supported"
+        );
+    }
+}
+
+#[test]
+fn build_file_current_record_preserves_file_metadata() {
+    let cfg = test_cfg();
+    let file = test_file_row(&cfg, "src/main.rs", "blob-1", 42, 420);
+    let record = build_file_current_record(
+        "src/main.rs",
+        "blob-1",
+        &file,
+        Some("Top-level docs".to_string()),
+    );
+
+    assert_eq!(record.symbol_id, file.symbol_id);
+    assert_eq!(record.artefact_id, file.artefact_id);
+    assert_eq!(record.canonical_kind.as_deref(), Some("file"));
+    assert_eq!(record.language_kind, "file");
+    assert_eq!(record.symbol_fqn, "src/main.rs");
+    assert_eq!(record.end_line, 42);
+    assert_eq!(record.end_byte, 420);
+    assert_eq!(record.docstring.as_deref(), Some("Top-level docs"));
+    assert_eq!(record.content_hash, "blob-1");
+}
+
+#[test]
+fn build_symbol_records_chain_file_and_nested_parent_links() {
+    let cfg = test_cfg();
+    let path = "src/ui.ts";
+    let blob_sha = "blob-ui";
+    let file = test_file_row(&cfg, path, blob_sha, 30, 300);
+    let items = vec![
+        JsTsArtefact {
+            canonical_kind: Some("class".to_string()),
+            language_kind: "class_declaration".to_string(),
+            name: "Widget".to_string(),
+            symbol_fqn: format!("{path}::Widget"),
+            parent_symbol_fqn: None,
+            start_line: 1,
+            end_line: 20,
+            start_byte: 0,
+            end_byte: 200,
+            signature: "export class Widget {}".to_string(),
+            modifiers: vec!["export".to_string()],
+            docstring: Some("Widget docs".to_string()),
+        },
+        JsTsArtefact {
+            canonical_kind: Some("method".to_string()),
+            language_kind: "method_definition".to_string(),
+            name: "render".to_string(),
+            symbol_fqn: format!("{path}::Widget::render"),
+            parent_symbol_fqn: Some(format!("{path}::Widget")),
+            start_line: 5,
+            end_line: 10,
+            start_byte: 40,
+            end_byte: 120,
+            signature: "render(): void {}".to_string(),
+            modifiers: vec![],
+            docstring: None,
+        },
+    ];
+
+    let records = build_symbol_records(&cfg, path, blob_sha, &file, &items);
+    assert_eq!(records.len(), 2);
+
+    let class_record = &records[0];
+    assert_eq!(class_record.parent_symbol_id, Some(file.symbol_id.clone()));
+    assert_eq!(
+        class_record.parent_artefact_id,
+        Some(file.artefact_id.clone())
+    );
+    assert_eq!(class_record.docstring.as_deref(), Some("Widget docs"));
+
+    let method_record = &records[1];
+    assert_eq!(
+        method_record.parent_symbol_id,
+        Some(class_record.symbol_id.clone())
+    );
+    assert_eq!(
+        method_record.parent_artefact_id,
+        Some(class_record.artefact_id.clone())
+    );
+    assert_eq!(
+        method_record.signature.as_deref(),
+        Some("render(): void {}")
+    );
+}
+
+#[test]
+fn build_historical_edge_records_keep_resolved_and_unresolved_targets() {
+    let cfg = test_cfg();
+    let path = "src/main.ts";
+    let blob_sha = "blob-2";
+    let from = test_symbol_record(&cfg, path, blob_sha, "from-symbol", "source", 1, 2);
+    let to = test_symbol_record(&cfg, path, blob_sha, "to-symbol", "target", 4, 5);
+    let current_by_fqn = [
+        (from.symbol_fqn.clone(), from.clone()),
+        (to.symbol_fqn.clone(), to.clone()),
+    ]
+    .into_iter()
+    .collect::<HashMap<_, _>>();
+
+    let records = build_historical_edge_records(
+        &cfg,
+        blob_sha,
+        "typescript",
+        vec![
+            test_call_edge(&from.symbol_fqn, &to.symbol_fqn, 7),
+            test_unresolved_call_edge(&from.symbol_fqn, "remote::symbol", 9),
+            test_call_edge("missing::from", &to.symbol_fqn, 11),
+        ],
+        &current_by_fqn,
+    );
+
+    assert_eq!(records.len(), 2);
+    assert_eq!(
+        records[0].to_symbol_id.as_deref(),
+        Some(to.symbol_id.as_str())
+    );
+    assert_eq!(
+        records[0].to_artefact_id.as_deref(),
+        Some(to.artefact_id.as_str())
+    );
+    assert!(records[0].to_symbol_ref.is_none());
+    assert!(records[1].to_symbol_id.is_none());
+    assert!(records[1].to_artefact_id.is_none());
+    assert_eq!(records[1].to_symbol_ref.as_deref(), Some("remote::symbol"));
+}
+
+#[test]
+fn build_current_edge_records_resolve_local_and_external_targets() {
+    let cfg = test_cfg();
+    let path = "src/main.ts";
+    let blob_sha = "blob-3";
+    let from = test_symbol_record(&cfg, path, blob_sha, "from-symbol", "source", 1, 2);
+    let to = test_symbol_record(&cfg, path, blob_sha, "to-symbol", "target", 4, 5);
+    let current_by_fqn = [
+        (from.symbol_fqn.clone(), from.clone()),
+        (to.symbol_fqn.clone(), to.clone()),
+    ]
+    .into_iter()
+    .collect::<HashMap<_, _>>();
+    let external_targets = [(
+        "pkg::remote".to_string(),
+        (
+            "external-symbol".to_string(),
+            "external-artefact".to_string(),
+        ),
+    )]
+    .into_iter()
+    .collect::<HashMap<_, _>>();
+
+    let records = build_current_edge_records(
+        &cfg,
+        "commit-3",
+        blob_sha,
+        "typescript",
+        vec![
+            test_call_edge(&from.symbol_fqn, &to.symbol_fqn, 7),
+            test_unresolved_call_edge(&from.symbol_fqn, "pkg::remote", 8),
+        ],
+        &current_by_fqn,
+        &external_targets,
+    );
+
+    assert_eq!(records.len(), 2);
+    assert_eq!(
+        records[0].to_symbol_id.as_deref(),
+        Some(to.symbol_id.as_str())
+    );
+    assert_eq!(
+        records[0].to_artefact_id.as_deref(),
+        Some(to.artefact_id.as_str())
+    );
+    assert_eq!(records[1].to_symbol_id.as_deref(), Some("external-symbol"));
+    assert_eq!(
+        records[1].to_artefact_id.as_deref(),
+        Some("external-artefact")
+    );
+    assert_eq!(records[1].to_symbol_ref.as_deref(), Some("pkg::remote"));
+}
+
+#[test]
+fn incoming_revision_is_newer_prefers_newer_timestamp_then_sha() {
+    assert!(incoming_revision_is_newer(None, "bbb", 10));
+    assert!(incoming_revision_is_newer(
+        Some(("aaa".to_string(), 9)),
+        "bbb",
+        10
+    ));
+    assert!(!incoming_revision_is_newer(
+        Some(("zzz".to_string(), 11)),
+        "bbb",
+        10
+    ));
+    assert!(incoming_revision_is_newer(
+        Some(("aaa".to_string(), 10)),
+        "bbb",
+        10
+    ));
+    assert!(!incoming_revision_is_newer(
+        Some(("ccc".to_string(), 10)),
+        "bbb",
+        10
+    ));
+}
+
+#[test]
+fn default_branch_name_uses_current_branch_and_falls_back_to_main() {
+    let repo = seed_git_repo();
+    git_ok(repo.path(), &["checkout", "-B", "feature/test-branch"]);
+
+    assert_eq!(default_branch_name(repo.path()), "feature/test-branch");
+    assert_eq!(
+        default_branch_name(tempdir().expect("temp dir").path()),
+        "main"
+    );
+}
+
+#[test]
+fn collect_checkpoint_commit_map_prefers_newest_db_mapped_checkpoint_commit() {
+    let repo = seed_git_repo();
+    let checkpoint_id = "aabbccddeeff";
+
+    git_ok(
+        repo.path(),
+        &["commit", "--allow-empty", "-m", "older checkpoint"],
+    );
+    let older_sha = git_ok(repo.path(), &["rev-parse", "HEAD"]);
+    git_ok(
+        repo.path(),
+        &["commit", "--allow-empty", "-m", "newest checkpoint"],
+    );
+    let newest_sha = git_ok(repo.path(), &["rev-parse", "HEAD"]);
+    insert_commit_checkpoint_mapping(repo.path(), &older_sha, checkpoint_id);
+    insert_commit_checkpoint_mapping(repo.path(), &newest_sha, checkpoint_id);
+
+    let checkpoint_map =
+        collect_checkpoint_commit_map(repo.path()).expect("checkpoint commit map should build");
+
+    assert_eq!(checkpoint_map.len(), 1);
+    let info = checkpoint_map
+        .get(checkpoint_id)
+        .expect("checkpoint should be present");
+    assert_eq!(info.subject, "newest checkpoint");
+    assert!(!info.commit_sha.is_empty());
+    assert!(info.commit_unix > 0);
+}
+
+#[test]
+fn collect_checkpoint_commit_map_reads_commit_checkpoints_table() {
+    let repo = seed_git_repo();
+    let checkpoint_id = "b0b1b2b3b4b5";
+
+    git_ok(
+        repo.path(),
+        &[
+            "commit",
+            "--allow-empty",
+            "-m",
+            "checkpoint without trailer",
+        ],
+    );
+    let commit_sha = git_ok(repo.path(), &["rev-parse", "HEAD"]);
+    insert_commit_checkpoint_mapping(repo.path(), &commit_sha, checkpoint_id);
+
+    let checkpoint_map =
+        collect_checkpoint_commit_map(repo.path()).expect("checkpoint commit map should build");
+
+    assert_eq!(checkpoint_map.len(), 1);
+    let info = checkpoint_map
+        .get(checkpoint_id)
+        .expect("checkpoint should be present");
+    assert_eq!(info.commit_sha, commit_sha);
+    assert_eq!(info.subject, "checkpoint without trailer");
+    assert!(info.commit_unix > 0);
+}
+
